# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Athlete {
  id: number;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  sport: string;
  experienceLevel: string;
  /** @nullable */
  weeklySchedule?: string | null;
  /** @nullable */
  goals?: string | null;
  onboardingComplete: boolean;
  createdAt: string;
}

export interface AthleteInput {
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  sport: string;
  experienceLevel: string;
  /** @nullable */
  weeklySchedule?: string | null;
  /** @nullable */
  goals?: string | null;
  onboardingComplete?: boolean;
}

export interface PersonalRecord {
  id: number;
  sport: string;
  category: string;
  value: string;
  unit: string;
  /** @nullable */
  notes?: string | null;
  achievedAt: string;
  createdAt: string;
}

export interface PersonalRecordInput {
  sport: string;
  category: string;
  value: string;
  unit: string;
  /** @nullable */
  notes?: string | null;
  achievedAt: string;
}

export interface CategorySummary {
  category: string;
  latestValue: string;
  unit: string;
  count: number;
}

export interface SportSummary {
  sport: string;
  count: number;
  categories: CategorySummary[];
}

export interface RecordsSummary {
  totalPRs: number;
  sports: SportSummary[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListRecordsParams = {
sport?: string;
category?: string;
};

export type GetRecentRecordsParams = {
limit?: number;
};