# lib/api-zod/src/generated/types/athlete.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface Athlete {
  id: number;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  sport: string;
  experienceLevel: string;
  /** @nullable */
  weeklySchedule?: string | null;
  /** @nullable */
  goals?: string | null;
  onboardingComplete: boolean;
  createdAt: Date;
}