# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  category: string;
  latestValue: string;
  unit: string;
  count: number;
}