# lib/api-zod/src/generated/types/personalRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface PersonalRecord {
  id: number;
  sport: string;
  category: string;
  value: string;
  unit: string;
  /** @nullable */
  notes?: string | null;
  achievedAt: Date;
  createdAt: Date;
}