# lib/api-zod/src/generated/types/recordsSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { SportSummary } from './sportSummary';

export interface RecordsSummary {
  totalPRs: number;
  sports: SportSummary[];
}