# lib/api-zod/src/generated/types/sportSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PR Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySummary } from './categorySummary';

export interface SportSummary {
  sport: string;
  count: number;
  categories: CategorySummary[];
}